+++ library-gen/Stratosphere/Resources.hs
@@ -23,6 +23,9 @@ module Stratosphere.Resources
      , resource
      , properties
      , deletionPolicy
+     , resCreationPolicy
+     , resUpdatePolicy
+     , dependsOn
      , ResourceProperties (..)
      , DeletionPolicy (..)
      , Resources (..)
@@ -40,6 +43,7 @@ import Stratosphere.Resources.DBSecurityGroupIngress as X
 import Stratosphere.Resources.Subnet as X
 import Stratosphere.Resources.DBInstance as X
 import Stratosphere.Resources.IAMRole as X
+import Stratosphere.Resources.LifecycleHook as X
 import Stratosphere.Resources.Group as X
 import Stratosphere.Resources.DBSubnetGroup as X
 import Stratosphere.Resources.SecurityGroup as X
@@ -54,6 +58,7 @@ import Stratosphere.Resources.VPCGatewayAttachment as X
 import Stratosphere.Resources.EIP as X
 import Stratosphere.Resources.User as X
 import Stratosphere.Resources.DBSecurityGroup as X
+import Stratosphere.Resources.LaunchConfiguration as X
 import Stratosphere.Resources.SubnetRouteTableAssociation as X
 import Stratosphere.Resources.RecordSetGroup as X
 import Stratosphere.Resources.Stack as X
@@ -61,6 +66,9 @@ import Stratosphere.Resources.ManagedPolicy as X
 import Stratosphere.Resources.VPC as X
 import Stratosphere.Resources.AccessKey as X
 import Stratosphere.Resources.LoadBalancer as X
+import Stratosphere.Resources.ScalingPolicy as X
+import Stratosphere.Resources.AutoScalingGroup as X
+import Stratosphere.Resources.ScheduledAction as X
 import Stratosphere.Resources.Volume as X
 import Stratosphere.Resources.UserToGroupAddition as X
 import Stratosphere.Resources.VPCEndpoint as X
@@ -73,12 +81,14 @@ import Stratosphere.ResourceProperties.ConnectionDrainingPolicy as X
 import Stratosphere.ResourceProperties.HealthCheck as X
 import Stratosphere.ResourceProperties.EC2SsmAssociationParameters as X
 import Stratosphere.ResourceProperties.ELBPolicy as X
+import Stratosphere.ResourceProperties.AutoScalingMetricsCollection as X
 import Stratosphere.ResourceProperties.LBCookieStickinessPolicy as X
 import Stratosphere.ResourceProperties.AliasTarget as X
 import Stratosphere.ResourceProperties.NetworkInterface as X
 import Stratosphere.ResourceProperties.SecurityGroupEgressRule as X
 import Stratosphere.ResourceProperties.EC2BlockDeviceMapping as X
 import Stratosphere.ResourceProperties.RecordSetGeoLocation as X
+import Stratosphere.ResourceProperties.AutoScalingBlockDeviceMapping as X
 import Stratosphere.ResourceProperties.PrivateIpAddressSpecification as X
 import Stratosphere.ResourceProperties.IAMPolicies as X
 import Stratosphere.ResourceProperties.ListenerProperty as X
@@ -88,10 +98,21 @@ import Stratosphere.ResourceProperties.EC2MountPoint as X
 import Stratosphere.ResourceProperties.SecurityGroupIngressRule as X
 import Stratosphere.ResourceProperties.RDSSecurityGroupRule as X
 import Stratosphere.ResourceProperties.EBSBlockDevice as X
+import Stratosphere.ResourceProperties.StepAdjustments as X
 import Stratosphere.ResourceProperties.AccessLoggingPolicy as X
+import Stratosphere.ResourceProperties.AutoScalingTags as X
+import Stratosphere.ResourceProperties.AutoScalingEBSBlockDevice as X
+import Stratosphere.ResourceProperties.AutoScalingNotificationConfigurations as X
 import Stratosphere.ResourceProperties.AppCookieStickinessPolicy as X
 import Stratosphere.ResourceProperties.ConnectionSettings as X
 import Stratosphere.ResourceProperties.UserLoginProfile as X
+
+import Stratosphere.ResourceAttributes.ResourceSignal as X
+import Stratosphere.ResourceAttributes.AutoScalingRollingUpdate as X
+import Stratosphere.ResourceAttributes.CreationPolicy as X
+import Stratosphere.ResourceAttributes.AutoScalingReplacingUpdate as X
+import Stratosphere.ResourceAttributes.UpdatePolicy as X
+import Stratosphere.ResourceAttributes.AutoScalingScheduledAction as X
 
 import Stratosphere.Helpers
 import Stratosphere.Values
@@ -101,6 +122,7 @@ data ResourceProperties
   | SubnetProperties Subnet
   | DBInstanceProperties DBInstance
   | IAMRoleProperties IAMRole
+  | LifecycleHookProperties LifecycleHook
   | GroupProperties Group
   | DBSubnetGroupProperties DBSubnetGroup
   | SecurityGroupProperties SecurityGroup
@@ -115,6 +137,7 @@ data ResourceProperties
   | EIPProperties EIP
   | UserProperties User
   | DBSecurityGroupProperties DBSecurityGroup
+  | LaunchConfigurationProperties LaunchConfiguration
   | SubnetRouteTableAssociationProperties SubnetRouteTableAssociation
   | RecordSetGroupProperties RecordSetGroup
   | StackProperties Stack
@@ -122,6 +145,9 @@ data ResourceProperties
   | VPCProperties VPC
   | AccessKeyProperties AccessKey
   | LoadBalancerProperties LoadBalancer
+  | ScalingPolicyProperties ScalingPolicy
+  | AutoScalingGroupProperties AutoScalingGroup
+  | ScheduledActionProperties ScheduledAction
   | VolumeProperties Volume
   | UserToGroupAdditionProperties UserToGroupAddition
   | VPCEndpointProperties VPCEndpoint
@@ -146,6 +172,9 @@ data Resource =
   { resourceName :: T.Text
   , resourceProperties :: ResourceProperties
   , resourceDeletionPolicy :: Maybe DeletionPolicy
+  , resourceResCreationPolicy :: Maybe CreationPolicy
+  , resourceResUpdatePolicy :: Maybe UpdatePolicy
+  , resourceDependsOn :: Maybe [Val T.Text]
   } deriving (Show)
 
 instance ToRef Resource b where
@@ -161,14 +190,21 @@ resource rn rp =
   { resourceName = rn
   , resourceProperties = rp
   , resourceDeletionPolicy = Nothing
+  , resourceResCreationPolicy = Nothing
+  , resourceResUpdatePolicy = Nothing
+  , resourceDependsOn = Nothing
   }
 
 $(makeFields ''Resource)
 
 resourceToJSON :: Resource -> Value
-resourceToJSON (Resource _ props dp) =
+resourceToJSON (Resource _ props dp cp up deps) =
     object $ resourcePropertiesJSON props ++ catMaybes
-    [ maybeField "DeletionPolicy" dp ]
+    [ maybeField "DeletionPolicy" dp
+    , maybeField "CreationPolicy" cp
+    , maybeField "UpdatePolicy" up
+    , maybeField "DependsOn" deps
+    ]
 
 resourcePropertiesJSON :: ResourceProperties -> [Pair]
 resourcePropertiesJSON (DBSecurityGroupIngressProperties x) =
@@ -179,6 +215,8 @@ resourcePropertiesJSON (DBInstanceProperties x) =
   [ "Type" .= ("AWS::RDS::DBInstance" :: String), "Properties" .= toJSON x]
 resourcePropertiesJSON (IAMRoleProperties x) =
   [ "Type" .= ("AWS::IAM::Role" :: String), "Properties" .= toJSON x]
+resourcePropertiesJSON (LifecycleHookProperties x) =
+  [ "Type" .= ("AWS::AutoScaling::LifecycleHook" :: String), "Properties" .= toJSON x]
 resourcePropertiesJSON (GroupProperties x) =
   [ "Type" .= ("AWS::IAM::Group" :: String), "Properties" .= toJSON x]
 resourcePropertiesJSON (DBSubnetGroupProperties x) =
@@ -207,6 +245,8 @@ resourcePropertiesJSON (UserProperties x) =
   [ "Type" .= ("AWS::IAM::User" :: String), "Properties" .= toJSON x]
 resourcePropertiesJSON (DBSecurityGroupProperties x) =
   [ "Type" .= ("AWS::RDS::DBSecurityGroup" :: String), "Properties" .= toJSON x]
+resourcePropertiesJSON (LaunchConfigurationProperties x) =
+  [ "Type" .= ("AWS::AutoScaling::LaunchConfiguration" :: String), "Properties" .= toJSON x]
 resourcePropertiesJSON (SubnetRouteTableAssociationProperties x) =
   [ "Type" .= ("AWS::EC2::SubnetRouteTableAssociation" :: String), "Properties" .= toJSON x]
 resourcePropertiesJSON (RecordSetGroupProperties x) =
@@ -221,6 +261,12 @@ resourcePropertiesJSON (AccessKeyProperties x) =
   [ "Type" .= ("AWS::IAM::AccessKey" :: String), "Properties" .= toJSON x]
 resourcePropertiesJSON (LoadBalancerProperties x) =
   [ "Type" .= ("AWS::ElasticLoadBalancing::LoadBalancer" :: String), "Properties" .= toJSON x]
+resourcePropertiesJSON (ScalingPolicyProperties x) =
+  [ "Type" .= ("AWS::AutoScaling::ScalingPolicy" :: String), "Properties" .= toJSON x]
+resourcePropertiesJSON (AutoScalingGroupProperties x) =
+  [ "Type" .= ("AWS::AutoScaling::AutoScalingGroup" :: String), "Properties" .= toJSON x]
+resourcePropertiesJSON (ScheduledActionProperties x) =
+  [ "Type" .= ("AWS::AutoScaling::ScheduledAction" :: String), "Properties" .= toJSON x]
 resourcePropertiesJSON (VolumeProperties x) =
   [ "Type" .= ("AWS::EC2::Volume" :: String), "Properties" .= toJSON x]
 resourcePropertiesJSON (UserToGroupAdditionProperties x) =
@@ -245,6 +291,7 @@ resourceFromJSON n o =
          "AWS::EC2::Subnet" -> SubnetProperties <$> (o .: "Properties")
          "AWS::RDS::DBInstance" -> DBInstanceProperties <$> (o .: "Properties")
          "AWS::IAM::Role" -> IAMRoleProperties <$> (o .: "Properties")
+         "AWS::AutoScaling::LifecycleHook" -> LifecycleHookProperties <$> (o .: "Properties")
          "AWS::IAM::Group" -> GroupProperties <$> (o .: "Properties")
          "AWS::RDS::DBSubnetGroup" -> DBSubnetGroupProperties <$> (o .: "Properties")
          "AWS::EC2::SecurityGroup" -> SecurityGroupProperties <$> (o .: "Properties")
@@ -259,6 +306,7 @@ resourceFromJSON n o =
          "AWS::EC2::EIP" -> EIPProperties <$> (o .: "Properties")
          "AWS::IAM::User" -> UserProperties <$> (o .: "Properties")
          "AWS::RDS::DBSecurityGroup" -> DBSecurityGroupProperties <$> (o .: "Properties")
+         "AWS::AutoScaling::LaunchConfiguration" -> LaunchConfigurationProperties <$> (o .: "Properties")
          "AWS::EC2::SubnetRouteTableAssociation" -> SubnetRouteTableAssociationProperties <$> (o .: "Properties")
          "AWS::Route53::RecordSetGroup" -> RecordSetGroupProperties <$> (o .: "Properties")
          "AWS::CloudFormation::Stack" -> StackProperties <$> (o .: "Properties")
@@ -266,6 +314,9 @@ resourceFromJSON n o =
          "AWS::EC2::VPC" -> VPCProperties <$> (o .: "Properties")
          "AWS::IAM::AccessKey" -> AccessKeyProperties <$> (o .: "Properties")
          "AWS::ElasticLoadBalancing::LoadBalancer" -> LoadBalancerProperties <$> (o .: "Properties")
+         "AWS::AutoScaling::ScalingPolicy" -> ScalingPolicyProperties <$> (o .: "Properties")
+         "AWS::AutoScaling::AutoScalingGroup" -> AutoScalingGroupProperties <$> (o .: "Properties")
+         "AWS::AutoScaling::ScheduledAction" -> ScheduledActionProperties <$> (o .: "Properties")
          "AWS::EC2::Volume" -> VolumeProperties <$> (o .: "Properties")
          "AWS::IAM::UserToGroupAddition" -> UserToGroupAdditionProperties <$> (o .: "Properties")
          "AWS::EC2::VPCEndpoint" -> VPCEndpointProperties <$> (o .: "Properties")
@@ -276,7 +327,10 @@ resourceFromJSON n o =
 
          _ -> fail $ "Unknown resource type: " ++ type'
        dp <- o .:? "DeletionPolicy"
-       return $ Resource n props dp
+       cp <- o .:? "CreationPolicy"
+       up <- o .:? "UpdatePolicy"
+       deps <- o .:? "DependsOn"
+       return $ Resource n props dp cp up deps
 
 -- | Wrapper around a list of 'Resources's to we can modify the aeson
 -- instances.