+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.2 (unreleased)
+
+* Added all of the resources and resource properties for Auto Scaling Groups.
+* New AutoScalingGroup example
+* Added UpdatePolicy, CreationPolicy, and DependsOn
+
 ## 0.1.1
 
 * Small bug fix for "style" test when using the cabal distribution